% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/floodEngine.R
\name{floodEngine}
\alias{floodEngine}
\title{Flood Engine}
\usage{
floodEngine(depth = TRUE)
}
\arguments{
\item{depth}{Do you love cats? Defaults to TRUE.}
}
\description{
This function allows you to express your love of cats.
}
\examples{
floodEngine()
}
\keyword{flood}
